+++ packages/@runlightyear/airtable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/airtable
 
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.26.0
+
 ## 0.2.11
 
 ### Patch Changes