+++ packages/@runlightyear/gcal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gcal
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.26.0
+
 ## 0.4.0
 
 ### Minor Changes