+++ packages/@runlightyear/gmail/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gmail
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.26.0
+
 ## 0.2.0
 
 ### Minor Changes