+++ packages/@runlightyear/notion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/notion
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.26.0
+
 ## 0.1.7
 
 ### Patch Changes