+++ packages/@runlightyear/postmark/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/postmark
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.26.0
+
 ## 0.1.11
 
 ### Patch Changes