+++ packages/@runlightyear/zoom/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/zoom
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.26.0
+
 ## 0.3.0
 
 ### Minor Changes